Made Board and Category classes based off of older build
