+++ ElementaryEducationMachine-master/src/main/java/edu/bsu/cs222/Jeopardy/Model/Category.java
@@ -1,0 +1,13 @@
+package edu.bsu.cs222.Jeopardy.Model;
+
+import java.util.HashMap;
+
+public class Category {
+    String categoryName;
+    HashMap<String, String> questions;
+
+    public Category(String categoryName, HashMap<String,String> questions){
+        this.categoryName = categoryName;
+        this.questions = questions;
+    }
+}
